Remove adding error listeners for most methods.
Only require addition of error listeners if we are handling events directly.
Most of the methods are simply calling an ssh/sftp method, passing in a callback
which will return an error if there is one i.e. the base function is already
handling the events, so no need for additional error handlers.