Relax error checking and change exists() to use stat()
Using list() in exists() causes issues if the user does not have read
access to parent directories. Using stat() should avoid this issue as
you wold need to be able to stat() the parent in order to create a file
or directory within it.

Also relaxe some error checks as the possible combinations for directories
and permissions combined with recursive mkdir create too many round trip calls,
which has a performance hit. Will just handle error if operation fails instead.
Downside is error message returned is not necessarily as informative as it
could be.