Move removeListener calls outside of finally block
Cannot do removeListenres in finally block of try/catch block because
the code being executed inside the block is asynchronous. The removeListener
will be called before the resolve/reject of the async block. Put removeListener
inside the async block so that it is only called after either resolve or reject
is called.