Change get() listeners. Use end not finish
Previously listened for finish event on destination stream to resolve
promise. However may not want or get finish signal from destination steam,
which would cause method to hang waiting on a promise which does not
resolve/reject. Change to listen for reader end event as that is emitted when
all data has been consumed (not just sent).