Cleanup connect code. Remove sleep call
At one stage, ssh2 was returning multiple errors when a connection failed. To
handle this, ssh2-sftp-client was doing a 1 second sleep after an error before
removing error listeners and then returning the first error (as that was the
most relevant error.

The ssh2 module has been updated to stop this behaviour, so have removed the
call to sleep and now just return the error which ssh2 throws.