Refactoring of event handlers and added some additional path checks
It appears that sftp servers running under Windows create a few issues which
are not apparent under Linux. In particular -

- Some error messages are next to useless (just reports 'Failure' with no
details).
- For some reason, some operations which would generate an error under Linux
seem to be silently ignored with Windows. Have added some additional tests to
deal with these situations.
- Windows seems to generate multiple error events for a single error. This is
particularly noticable with connection failures. To make it worse, the second
error seems to occur after a small delay. As a result, the module only catches
the first error and then crashes with an uncaught exception error. To work
around this, I have added a short 500ms delay after a connection fails. This
seems to fix the issue, but at the cost of making connections slightly slower.