Modify event used to resolve get() requests
Previously, get() resolves when reader emits an end event. However, in some
circumstances this can cause errors if client code immediately attempts to
access the retrieved file as it is possible the last chunk has not completed
being flushed to the file. This is particularly noticeable on slower
destinations (like a cloud bucket). However, we cannot use the close event on
the write stream as sometimes clients want to keep the stream open (to append
more data).

This change now checks the pipe options. If the pipe options set end to
false (defualt is true), pipe() does not call close on the write stream. If this
is the case, we still listen for the end() even on the read stream, but if not,
we use the close event on the write stream to resolve the get() promise.