Refactor of event listener handling to remove global data structure
The previous management of event listeners used a hash indexed by method name
to track event handlers so they could be removed when the method completes.
However, this did not work well with recursive or parallel execution of methods.

New version now closes over an object storing the end, close and error handler
so that each invocation of the method has its own private copy.