Modified rmdir to perform file deletion in parallel
As rmdir can be slow if the tree is large, this change attempts to improve
performance by deleting files within a directory in parallel. Recursive deletion
of directories is still performed synchronously because it is too easy to blow
out with large trees. Even with parallel deletion of files, there is still a
risk of memory blow out for directories with very large numbers of files.

Will probably add a switch to turn off parallel deletion of files to handle
those edge cases where there are too many files to cope.