Modify uploadDir and downloadDir to accept a filter function
Previously, these methods accepted a regexp/glob to filter, but that was not
flexible enough. Now these methods accept a function which will be called with
two arguments. The first argument is the full path to the item to be
uploaded/downloaded and the second argument is a boolean which will be tgrue if
the path represents a directory.