Throw exception when rmdir target does not exist
Previously, if you tried to remove a directory which did not exist, we just
resolved the promise and returned with no error. This could result in subtle
client errors being missed. Now throws a Bad path error with a No such file
message if the target directory passed to rmdir does not exist.