Fix timing bug in get() method
When using get() without a destination argument, the method
was returning the concat-stream object rather than the data buffer.
This was due to a timing issue where the end event was firing before the call
to resolve() in the concat-stream function. As a result, the concat stream was
being returned instead of the buffer which is contained inside that object.