Make default close and end listeners simpler
The default end and close listeners were trying to be more clever than is
absolutely necessary by trying to assess if an  end/close event was an expected
event being fired as the result of expected/requested actions i.e. ending the connection
and end/close events raised when not actually expected. However, this strategy relies too
heavily on the events being fired closely to the 'parent' or triggering
event/action. Sometimes, the dealy would be sufficient to lose the connection to the
'thing' which triggered the end/close and would subsequently be seen as unexpected.

In reality, we don't really care. All that matters is that if an end or close event does
fire, we need to invalidate the sftp connection. Alerting the client that an event which
may not have been expected has occurred is not of much use. So now, we log that a
close/end event was fired and we invalidate the sftp connection and leave it at that.