Fix destroying of reader stream in put()
GH Issue #543
When a write error occurred, the read stream was not destroyed.
This could result in too many file handles being left open. Note
that the fix only ensures a read stream created inside the method is
destroyed. When a read stream is passed into the method, it is the
responsibility of the calling code to manage stream cleanup.