Rework local and global event handling
While the existing handling of events worked well with well behaving
servers that sent an error before unexpectedly closing a connection,
there are some buggy/non-compliant servers out there which will just end a connection
without any error event or error information. Therefore I have restored the old behaviour
where unexpected end or close events raise an error. The danger here is that it is
possible that when we have a well behaving server, an end event may get handled before the
error event This will result in a non-specific error being raised and not providing any
useful diagnostic details about the error. In worst case situations, it is also possible
that the error event will cause a second error to be thrown which will not be caught by
this module, bubbling up and crashing the procewss with an uncaught error exception.