Move back to CommonJS module
While ESM modules are the direction things are moving, found that moving this module to
ESM created far more problems for CommonJS users than having it as a CommonJS module which
EMS module users have to adapt for. Maybe re-evaluate this down the track and once more
tooling has matured etc.