Initial re-working of global event handlers
Event handlers which throw errors are a real problem for client code as they cannot catch
these errors. During active interaction with the sftp connection the problem does not
exist because the interaction is wrapped in a promise and the temporary listeners added
during that interaction have access to the promise reject function. However, because we
are dealing with a network connection, external events can destroy that connection outside
actibve interaction e.g. pull the network cable. We cannot just throw an error because
that will happen in the wrong execution context and cannot bwe caught by client code.

Solution is to allow client code to pass in callbacks which will be called when an error,
end or close event occurs OUTSIDE active interaction with the connection i.e. when we are
not calling one of the API methods. In addition, default callback functions are in place
which will just print info to the console.