Change some stream error handlers from 'once' to 'on' handlers.
Issue #579. If more an one error event was triggered by a stream, only the first one would
be  handled because the handler was defined with 'once'. This would result in the
subsequent error event not being handled, bubbling up to the top level and causing the
parent node process to crash with an unhandled error exception.

Note that the global event handlers were not handling this event because the flag used to
indicate the error wsas handled was still incorrectly set to handled from the 1st error
event triggered i.e. global error handlers don't work when multiple errors are triggered
by a single API call.